maintenance_system/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
binaryornot==0.4.4
certifi==2024.2.2
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cookiecutter==2.6.0
debugpy==1.8.1
decorator==5.1.1
Django==5.0.2
django-browser-reload==1.12.1
django-crispy-forms==2.1
django-nextjs==2.4.0
django-ninja==1.1.0
django-tailwind==3.8.0
django-widget-tweaks==1.5.0
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
idna==3.6
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rich==13.7.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
text-unidecode==1.3
tornado==6.4
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
